buffer limit needs to be 64 bit

+++ src/datasources/ascii/asciisourceconfig.h
@@ -105,7 +105,7 @@ class AsciiSourceConfig {
     NamedParameter<int, Key_unitsLine, Tag_unitsLine> _unitsLine;
     NamedParameter<bool, Key_useDot, Tag_useDot> _useDot;
     NamedParameter<bool, Key_limitFileBuffer, Tag_limitFileBuffer> _limitFileBuffer;
-    NamedParameter<int, Key_limitFileBufferSize, Tag_limitFileBufferSize> _limitFileBufferSize;
+    NamedParameter<qint64, Key_limitFileBufferSize, Tag_limitFileBufferSize> _limitFileBufferSize;
     NamedParameter<int, Key_useThreads, Tag_useThreads> _useThreads;
     NamedParameter<double, Key_dataRate, Tag_dataRate> _dataRate;
     NamedParameter<bool, Key_offsetDateTime, Tag_offsetDateTime> _offsetDateTime;
